feat: fixing move table height
